requirements.txt
Django==3.2.16
asgiref==3.4.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
defusedxml==0.7.1
django-allauth==0.52.0
django-cors-headers==3.8.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
idna==3.4
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2021.1
requests-oauthlib==1.3.1
six==1.16.0
typing-extensions==4.4.0
urllib3==1.26.14
autopep8==2.0.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
Django==3.2.16
django-ckeditor==6.3.2
django-crispy-forms==1.14.0
django-js-asset==2.0.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
Pillow==9.3.0
pycodestyle==2.10.0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
webencodings==0.5.1
whitenoise==6.2.0
